chore: add translation and remove unimported files
