+++ package.json
@@ -45,9 +45,7 @@
     "@emotion/react": "11.4.0",
     "@emotion/server": "^11.4.0",
     "@emotion/styled": "11.3.0",
-    "@popperjs/core": "^2.9.2",
     "abcq": "1.0.2",
-    "framer-motion": "4.1.17",
     "graphql": "15.5.0",
     "next": "10.2.3",
     "next-auth": "3.23.3",
@@ -56,12 +54,8 @@
     "react-dom": "17.0.2",
     "react-hook-form": "7.6.8",
     "react-lottie": "1.2.3",
-    "react-popper": "2.2.5",
-    "react-use": "17.2.4",
     "tslib": "2.2.0",
-    "use-dark-mode": "2.3.1",
-    "use-debounce": "6.0.1",
-    "uuid": "8.3.2"
+    "use-dark-mode": "2.3.1"
   },
   "devDependencies": {
     "@babel/cli": "^7.14.3",
