+++ public/static/locales/de/wishlist.json
@@ -1,0 +1,6 @@
+{
+  "headline": "Sag uns was du von einer Präsentations Software erwartest",
+  "button": {
+    "wish": "Ich habe einen Wunsch"
+  }
+}
