+++ public/static/locales/en/wishlist.json
@@ -1,0 +1,6 @@
+{
+  "headline": "Tell us what you expect from a persentation software",
+  "button": {
+    "wish": "I have a wish"
+  }
+}
