--- src/atoms/data-log/styled.tsx
@@ -1,10 +1,0 @@
-import styled from "@emotion/styled";
-
-export const StyledPre = styled.pre`
-	width: min-content;
-	margin: 1em;
-	padding: 1em;
-	border-radius: 2px;
-	background: #333;
-	color: #fec;
-`;
