--- src/atoms/icon/styled.ts
@@ -1,9 +1,0 @@
-import styled from "@emotion/styled";
-
-export const StyledSvg = styled.svg`
-	fill: currentColor;
-`;
-
-export const StyledPath = styled.path`
-	fill: currentColor;
-`;
