--- src/ions/constants/index.ts
@@ -1,1 +1,0 @@
-export const ROOT = "ROOT";
