--- src/ions/contexts/render.ts
@@ -1,4 +1,0 @@
-import { Renderer } from "@/ions/enums";
-import React from "react";
-
-export const RenderContext = React.createContext({ renderer: Renderer.html });
