--- src/ions/hooks/position.ts
@@ -1,4 +1,0 @@
-import { PositionContext } from "@/ions/contexts/position";
-import React from "react";
-
-export const usePositionContext = () => React.useContext(PositionContext);
