--- src/ions/icons/index.ts
@@ -1,3 +1,0 @@
-import { icons } from "./path-d";
-
-export { icons };
