--- src/ions/icons/path-d.d.ts
@@ -1,4 +1,0 @@
-import { Icons } from "@/types";
-
-export declare const icons: Icons;
-export default icons;
