--- src/organisms/header/styled.ts
@@ -1,3 +1,0 @@
-import styled from "@emotion/styled";
-
-export const StyledHeader = styled.header``;
