--- src/organisms/header/types.ts
@@ -1,3 +1,0 @@
-import { BasePropsWithInnerRef } from "@/types/components";
-
-export interface HeaderProps extends BasePropsWithInnerRef {}
