+++ src/pages/api/auth/[...nextauth]/index.tsx
@@ -1,7 +1,7 @@
-import NextAuth from "next-auth";
-import Providers from "next-auth/providers";
-import Adapters from "next-auth/adapters";
 import { PrismaClient } from "@dekk-app/dekk-backend/src/colonies/prisma/client";
+import NextAuth from "next-auth";
+import Adapters from "next-auth/adapters";
+import Providers from "next-auth/providers";
 
 const prisma = new PrismaClient();
